# Conflicts:
#	build.gradle
#	src/main/java/com/codegym/aurora/controller/back_office/LoginController.java
#	src/main/java/com/codegym/aurora/controller/back_office/UserController.java
#	src/main/java/com/codegym/aurora/repository/UserRepository.java
#	src/main/java/com/codegym/aurora/service/UserService.java
#	src/main/java/com/codegym/aurora/service/impl/UserServiceImpl.java

+++ src/main/java/com/codegym/aurora/repository/UserDetailRepository.java
@@ -4,5 +4,5 @@ import com.codegym.aurora.entity.UserDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface UserDetailRepository extends JpaRepository<UserDetail, Integer> {
+public interface UserDetailRepository extends JpaRepository<UserDetail, Long> {
 }
